# openapi.yaml
openapi: 3.0.3
info:
  title: Portal de Benefícios KNN API
  description: |
    API para o Portal de Benefícios KNN, permitindo que alunos da KNN Idiomas acessem promoções
    exclusivas de parceiros comerciais através de códigos de validação temporários.
  version: 1.0.0
  contact:
    name: Equipe de TI KNN Idiomas
    email: ti@knnidiomas.com.br
servers:
  - url: https://api.knnidiomas.com.br/v1
    description: Ambiente de produção
  - url: http://localhost:8000/v1
    description: Ambiente de desenvolvimento local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT com claim 'role' (student|partner|admin)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_CODE
            msg:
              type: string
              example: Código expirado ou já usado
          required:
            - code
            - msg
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
          description: Dados da resposta (varia conforme endpoint)
        msg:
          type: string
          default: ok
      required:
        - data
        - msg
...
    Partner:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trade_name:
          type: string
          example: Livraria Cultura
        category:
          type: string
          example: Livraria
        address:
          type: string
          example: Av. Paulista, 2073 - São Paulo/SP
        active:
          type: boolean
          default: true
      required:
        - id
        - trade_name
        - category
        - address
        - active
...
    Promotion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        title:
          type: string
          example: 20% de desconto em livros de idiomas
        type:
          type: string
          example: discount
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
      required:
        - id
        - partner_id
        - title
        - type
        - valid_from
        - valid_to
        - active
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cpf_hash:
          type: string
          format: sha256
        name:
          type: string
          example: João da Silva
        email:
          type: string
          format: email
        course:
          type: string
          example: Inglês Avançado
        active_until:
          type: string
          format: date
      required:
        - id
        - cpf_hash
        - name
        - email
        - course
        - active_until
    ValidationCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        code:
          type: string
          pattern: '^[0-9]{6}$'
          example: '123456'
        expires:
          type: string
          format: date-time
        used_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - student_id
        - partner_id
        - code
        - expires
    Redemption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        validation_code_id:
          type: string
          format: uuid
        value:
          type: number
          format: float
          example: 50.00
        used_at:
          type: string
          format: date-time
      required:
        - id
        - validation_code_id
        - value
        - used_at
...
    NotificationRequest:
      type: object
      properties:
        target:
          type: string
          enum: [students, partners]
        filter:
          type: object
          description: Filtros para selecionar destinatários
        message:
          type: string
        type:
          type: string
          enum: [email, push, both]
          default: both
      required:
        - target
        - message
...
  parameters:
...
    PaginationLimit:
      name: limit
      in: query
      description: Número máximo de itens por página
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    PaginationOffset:
      name: offset
      in: query
      description: Offset para paginação
      schema:
        type: integer
        default: 0
        minimum: 0
...
    EntityParam:
      name: entity
      in: path
      description: Tipo de entidade
      required: true
      schema:
        type: string
        enum: [students, partners, promotions, validation_codes, redemptions]
    IdParam:
      name: id
      in: path
      description: ID do recurso
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - bearerAuth: []
paths:
...
  /admin/{entity}:
    get:
      summary: Lista entidades
      description: Retorna lista de entidades (students, partners, promotions, etc.)
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/EntityParam'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Lista de entidades
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              oneOf:
                                - $ref: '#/components/schemas/Student'
                                - $ref: '#/components/schemas/Partner'
                                - $ref: '#/components/schemas/Promotion'
                                - $ref: '#/components/schemas/ValidationCode'
                                - $ref: '#/components/schemas/Redemption'
                          total:
                            type: integer
                          limit:
                            type: integer
                          offset:
                            type: integer
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entidade não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Cria nova entidade
      description: Cria uma nova entidade (student, partner, promotion, etc.)
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/EntityParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Student'
                - $ref: '#/components/schemas/Partner'
                - $ref: '#/components/schemas/Promotion'
      responses:
        '201':
          description: Entidade criada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/Student'
                          - $ref: '#/components/schemas/Partner'
                          - $ref: '#/components/schemas/Promotion'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entidade não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/{entity}/{id}:
    put:
      summary: Atualiza entidade
      description: Atualiza uma entidade existente (student, partner, promotion, etc.)
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/EntityParam'
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Student'
                - $ref: '#/components/schemas/Partner'
                - $ref: '#/components/schemas/Promotion'
      responses:
        '200':
          description: Entidade atualizada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/Student'
                          - $ref: '#/components/schemas/Partner'
                          - $ref: '#/components/schemas/Promotion'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entidade não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Remove/inativa entidade
      description: Remove ou inativa uma entidade existente (student, partner, promotion, etc.)
      tags:
        - admin
      parameters:
        - $ref: '#/components/parameters/EntityParam'
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Entidade removida/inativada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entidade não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/notifications:
    post:
      summary: Envia notificações
      description: Envia notificações push/e-mail para alunos ou parceiros
      tags:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      responses:
        '202':
          description: Notificações enfileiradas com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recipients:
                            type: integer
                            description: Número de destinatários
                          message_id:
                            type: string
                            format: uuid
                            description: ID da mensagem para rastreamento
                        required:
                          - recipients
                          - message_id
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'